Hi,
You can send me a pull request, and if the feature is good I will consider to merge it to the develop version. Thanks.

https://github.com/wenzhixin/bootstrap-table/pulls

--
Best regards,
Zhixin Wen
China

2015-02-26 21:14 GMT+08:00 Yuanyan Wu <ywu999@hotmail.com>:
Hi, Wen Zhi Xin,
 
Your script was very helpful for me. I found out an extra feature that could be added into it. Basically I wanted to sort a column with a customizable sort-name, not the default sort-name in the header, or the field of the column. This feature is very useful for the following scenario:
 
I have a column field “name” with the value of abc, but I want to display in html (field name: htmlName) as “<b><span style=”color:red”>abc</span></b>”. I really want to sort the column use the field “name” instead of “htmlName”.
 
If you are willing, I can commit my changes to your project.
 
Thanks,
 
Yuanyan Wu