Added logic for onClickCell and onDblClickCell to initBody
This allows for developers to listen for cell click and cell dblclick
events. The logic gets the cellIndex from the target <td>, then
selects corresponding <th> from this.$header using the "eq()"
selector. Using the 'field' from the header cell's data, the value is
selected from the data to be passed to the new events.

The key changes are as follow:
- 'click-cell' is fired at the same time as 'click-row'
- 'dbl-click-cell' is fired at the same time as 'dbl-click-row'
- changed selector of the dblclick event to "> tr > td"