Check length when comparing in refreshOptions
If we don't check length in compareObjects, we won't notice when the new
options object has had a new property added. (Because compareObjects
otherways only checks that everything present in the old options exist
in the new, not the other way round.) This might be especially relevant
when you want to activate an extension like reorderableRows.

Really, you should probably just remove that switch from the method
entirely
because it makes no sense to not check lengths first when comparing two
objects,
but I'm assuming there must be a reason for it.