Remove extra db call to server when possible
Bootstrap table does a db call for initial data to the server. This
causes the client to make two db calls. one when the client initializes(
gets default sorting), and another when filtering the database based on
the cookies that are saved on the browser. this change removes that
extra initial db call under three
conditions:

1) the cookies option is set to true
2) there are cookies associated to this table
3) all select dropdown filters have default values

this change also updates the documentation for filterControl, in order
to add the proper usage for the property filterData.